test/auth: extract certificate code
Not a functional change, just cleaning up
